requirements.txt
selenium
lxml
requests
python-dotenv
pyautogui